Update check_excels.py
In the previous version we already had colors gradient, in this commit I added function to check the numbers regardless of their decimal location and aim is to look for shifting in numbers(1.2345, 12.345, 0.12345) I also add a new function to change colors from green to blue to better distinguish the cells.
Also I added a counter for unique duplicates so if we see color changes based on the number of unique duplicates. It means with less number of duplicates the difference between them is bolder.